remove strict option from mypy (#38)
This PR proposes to remove the `--strict` option from the mypy check.

At the current state of the repo, it slows things down significantly,
and it is not clear if the benefits are that important for now. Of
course, we want a stable monitoring tool, and eventually we should clean
up the code and do proper type annotations, but now it seems it will
take a very long time to have something working reasonably, partly
because the --`strict` makes it much harder to merge PR #18, which adds
fundamental functionality to the tool.