Significant restructuring of repo (#46)
With this PR, we restructure the repo by introducing a base class named
"TemplateTest", which contains some useful functions, and the intent of
which is to have every new test being a subclass of that class.

The PR has been tested as it was applied and it looks quite stable, with
both tests running (in particular, the Quasimodo calls succeed). To
reduce noise in the logs, the quasimodo test is currently commented out.