quote test ignores 'NoLiquidity' and 'SellAmountDoesNotCoverFees' errors
The hash 0xda857a0db563dae564b09febb683fff6150628c1706afc9ebd961c194ca29c5e
contains a partially-fillable order with an executed amount which is too small
to pay for fees. This resulted in the hash being rechecked over and over.
This error is now handled by logging an error and letting the test succeed.