Remove faulty absolute tolerance from fee quote test (#56)
The fee test claims to use an absolute tolerance in ETH. It does not,
however, do this since there is no conversion of the token amount to
ETH.

I completely removed the absolute check here. The data for converting to
ETH is easily available and since fees should be more or less of the
same order of magnitude, a relative test should be enough.