Reduce amount of alerts (#60)
This PR reduces the amount of all three types of alerts.

The tolerances for all tests are increased. The values are chosen such
that there are about 1-4 alerts per day per test (based on log data from
the last 2 days).

This should make it managable to check all auctions producing an alert.

At the same time, tolerances for logging are reduced to make it easier
to inspect aggregate data in our logs.