[WIP] Add reference solver test (#58)
This PR adds another test which uses a reference solver. It is still
work in progress.

The reference solver (mostly quasimodo) solves simplified auction
instances for each order included in a settlement. If the reference
solver provides more surplus a log/alert is emitted.

The test works similarly to how the competition surplus test works. The
main difference is that alternative executions of orders are not taken
from the solution of other solvers but from quasimodo.

Two new APIs are introduced:
- An API for calling a reference solver
- An API for getting auction instance files from AWS.