Change surplus based tests to use new data from orderbook (#91)
This PR addresses #90 for the competition surplus test and combinatorial
auction test.

The main changes are:
- Order information is fetched from the orderbook orders endpoint.
- Execution data is obtained via the new data in the solver competition
endpoint.
- Fees and costs are not explicitly part of the test anymore.

Before, information on trades (order data and order execution) were read
from calldata. For non-winning solutions this is not possible anymore.
For the winning solution it is still possible. But due to changes to the
ordering of orders in the competition data, it was not easy for me to
reconstruct the link from uid to trade. Thus, all information is not
obtained through the orderbook api.

In principle, information on fees for non-winning solutions could be
computed from clearing prices in combination with effective buy and sell
amount. This is not done here since fees do not matter for surplus. Also
clearing prices are not yet part of the data in the competition
endpoint.
Information on costs for non-winning solutions is and will not
available. A test of the form `fees >= costs` had to be removed due to
that.

The old code had a "bug" where only one execution per solver was
studied. This is fixed now and the test should work correctly with
solvers submitting multiple solutions. The fix required changing the
data format for alternative solutions from a `dict` to a `list`