Adding tolerance to the combinatorial auction alert (#108)
This PR adds a tolerance to the alert of the combinatorial auction test.

This is implemented by adding information on surplus differences to
`filter_mask`. That information is then checked later and compared to
the new constant `COMBINATORIAL_AUCTION_ABSOLUTE_DEVIATION_ETH`. That
constant is for now set to `0.001 ETH`.

For logging purposes, the function `convert_fractions_to_floats` was
extended to also handle lists.

The code was tested on a recent alert for the transaction hash
`"0x1541857aaf67c8d06028a92202c8b44a731c51bcd2819d1a0a62faf00f073227"`.
There, baseline solver was providing `0.0004 ETH` more surplus. Before
the change, the test emits an alert. After the change, the test only
logs an info message.

This PR closes #107.