Add test for CoW AMM commitment (#109)
With enabling phase 2 of CoW AMM, solvers can execute CoW AMM orders
with custom amounts. This requires commiting and uncommiting the order
in the CoW AMM smart contract. Not uncommiting the order is a violation
of social consensus rules since it prevents the default order from
getting executed.

This PR implements a test where after every settlement which contains a
precommit interaction from the CoW AMM address the commited order for
that address is checked.

Since there seem to be no commits so far, the code is a bit difficult to
test. I added a test with tailored call data.

---------

Co-authored-by: harisang <harisangelidakis@gmail.com>