Increase threshold for combinatorial auction test (#111)
At the moment, the combinatorial auction test generates non-actionable
alerts.

The original threshold of `0.001 ETH` is not large enough to avoid false
positives due to, e.g., inaccurate fee estimates. The new threshold of
`0.005 ETH` is more in line with other thresholds used.

Midterm we could also remove the test or redesign it to have less
overlap with the competition surplus test.