Remove tests (#115)
This PR removes tests.

There have been problems with long delays in EBBO tests. This is
probably due to the number of tests run and their efficiency. To see
where the issues are coming from and for general clean up, some tests
are removed from the daemon.

- Reference solver surplus test: does not work at the moment.
- Buffer monitoring test: not related to EBBO and seems to slow down the
loop significantly
- CoW AMM commitment test: obsolete with the new version of CoW AMMs

For now the code for tests is not removed completely as they can still
be run manually. Longterm, all of the code should be removed as well.