Add high score alert (#122)
This PR adds a very simple alerts that checks if the winning score for
auctions that made it onchain is larger than some threshold. Very large
scores might be an indication of bogus native prices so this is one way
to detect those cases.

I realized that currently the tests, as they are set up, will pass if
the wrapper function returns True. But this doesn't really check the
logic the test implements, so in order to properly see logs and
behavior, one needs to run the tests as follows. E.g:

`python -m pytest --log-cli-level=DEBUG  tests/e2e/high_score_test.py`