Fix getting transaction hash from competition data (#123)
This PR changes how transaction hashes are fetched from competition
data.

To support multiple winners, the competition data format was changed to
allow for multiple transactions to be associated to an auction. This
meant, that instead of a `"transactionHash"` field containing a string,
a `"transactionHashes"` field containing a vector of strings is stored.
This PR, changes the code to use the new convention.